In the Directory Page, add all contact details from this page - https://bsrtc.co.in/schedules- including the division officer's name and other details. (Also keep all the text Left aligned)
Gallery - Let's upload the images that I have sent to you of Buses (On drive).

Remove polices from Contact us page.

In footer, Home, My Account (Lead to sign in page) and FAQ pages is hyperlink not as headlines.

In About Us Page - Reduce the gap between photos and VISION